% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{cdf}
\alias{cdf}
\title{Cumulative Distribution Function}
\usage{
cdf(object, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A numeric vector of cumulative probabilities for the supplied
quantiles, as returned by the dispatched method.
}
\description{
Generic function for computing cumulative distribution function.
}
\examples{
pn <- ProductNormal(mu = c(0.5, 0.3), Sigma = diag(2) * 0.1)
cdf(pn, q = 0)
}
