% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chi2}
\alias{chi2}
\title{Computes asymptotic and/or parametric bootstrap \eqn{\chi^2} test of a general function of direct and indirect effects}
\usage{
chi2(model, data, constraints = "a*b==0", alpha = 0.05, boot = "none",
  R = 200, parallel = "none", ncpus = 2L, ...)
}
\arguments{
\item{model}{a \link{lavaan} \link{model.syntax}. The \link{model.syntax} may not include constraints as defined in \code{constraints}}

\item{data}{a data frame  containing the observed data from the sample.}

\item{constraints}{a general function of the model parameters stated in terms of (in)equality constraint.}

\item{alpha}{significance level. Default is .05.}

\item{boot}{if \code{boot="none"} (default), asymptotic \eqn{\chi^2} test is performed. If \code{boot="parametric"}, parametric \eqn{\chi^2} test is produced.}

\item{R}{number of bootstrap samples.}

\item{parallel}{The type of parallel operation to be used (if any). If missing, the default is "no".}

\item{ncpus}{Integer: number of processes to be used in parallel operation. Default is 2L.}

\item{...}{other options. See \link{sem}.}
}
\value{
a \link{list} of the following objects:
\item{LRT}{Asymptotic likelihood ratio test. One can also extract an asymptotic p value here. See \link{anova} for more detail.}
\item{p}{bootstrap p value for the \eqn{\chi^2}. This occurs only when \code{boot="parametric"}}
\item{fitH0}{Model estimate for the more restricted model. This is a \link{lavaan} object. One can extract all the information from this using \link{lavaan} functions.}
}
\description{
Computes asymptotic and/or parametric bootstrap \eqn{\chi^2} test of a general function of direct and indirect effects
}
\examples{
data(PoliticalDemocracy)
m1 <- '
ind60 =~ x1 + x2 + x3
dem60 =~ y1 + y2 + y3 + y4
dem65 =~ y5 + y6 + y7 + y8
dem60 ~ a*ind60
dem65 ~ cp*ind60 + b* dem60
y1 ~~ y5
y2 ~~ y4 + y6
y3 ~~ y7
y4 ~~ y8
y6 ~~ y8
'
# Fit the mode without estimation
( chi2Test <- chi2(m1, data = PoliticalDemocracy ,"a*b==0", boot="parametric", R=20 , parallel = "multicore", ncpus=2) )
}

