% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_s7_compute.R
\name{.compute_cdf_dop}
\alias{.compute_cdf_dop}
\title{Internal: Compute CDF using Distribution of Product}
\usage{
.compute_cdf_dop(object, q)
}
\arguments{
\item{object}{ProductNormal object}

\item{q}{Quantile value}
}
\value{
Probability \eqn{P(XY \le q)}
}
\description{
Computes the cumulative distribution function \eqn{P(XY \le q)} for the product
of two normal random variables using numerical integration (Meeker & Escobar, 1994).
}
\keyword{internal}
