% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_s7_compute.R
\name{.compute_cdf_mc}
\alias{.compute_cdf_mc}
\title{Internal: Compute CDF using Monte Carlo}
\usage{
.compute_cdf_mc(object, q, n.mc = 1e+05)
}
\arguments{
\item{object}{ProductNormal object}

\item{q}{Quantile value}

\item{n.mc}{Monte Carlo sample size}
}
\value{
Probability \eqn{P(XY \le q)}
}
\description{
Computes the cumulative distribution function \eqn{P(XY \le q)} for the product
of normal random variables using Monte Carlo simulation. Works for any number
of variables.
}
\keyword{internal}
