% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_s7_compute.R
\name{.compute_quantile_dop}
\alias{.compute_quantile_dop}
\title{Internal: Compute quantile using Distribution of Product}
\usage{
.compute_quantile_dop(object, p, lower.tail = TRUE)
}
\arguments{
\item{object}{ProductNormal object}

\item{p}{Probability}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X <= x)}
}
\value{
Quantile value
}
\description{
Internal: Compute quantile using Distribution of Product
}
\keyword{internal}
