% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_s7_compute.R
\name{.compute_quantile_dop_internal}
\alias{.compute_quantile_dop_internal}
\title{Internal: Compute quantile using Meeker & Escobar (1994) algorithm}
\usage{
.compute_quantile_dop_internal(
  p,
  mu.x,
  mu.y,
  se.x,
  se.y,
  rho = 0,
  lower.tail = TRUE
)
}
\description{
Internal: Compute quantile using Meeker & Escobar (1994) algorithm
}
\keyword{internal}
