% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_s7_compute.R
\name{.compute_quantile_mc}
\alias{.compute_quantile_mc}
\title{Internal: Compute quantile using Monte Carlo}
\usage{
.compute_quantile_mc(object, p, n.mc = 1e+05, lower.tail = TRUE)
}
\arguments{
\item{object}{ProductNormal object}

\item{p}{Probability}

\item{n.mc}{Monte Carlo sample size}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X <= x)}
}
\value{
Quantile value
}
\description{
Internal: Compute quantile using Monte Carlo
}
\keyword{internal}
