% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Structural Equation Model}
\usage{
fit_model(model, data)
}
\arguments{
\item{model}{A character string representing \code{lavaan} model syntax, a \code{lavaan} model object, or an \code{OpenMx::MxModel} object.}

\item{data}{A data frame containing the data to which the model will be fitted.}
}
\value{
A fitted model object from either \code{lavaan} or \code{OpenMx}, depending on the input.
}
\description{
Fits a structural equation model to provided data using either the \code{lavaan} or \code{OpenMx} package.
}
\examples{
\dontrun{
data("HolzingerSwineford1939", package = "lavaan")
lav_model_syntax <- "visual =~ x1 + x2 + x3"
fitted_model <- fit_model(lav_model_syntax, HolzingerSwineford1939)
}
}
