% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemImputedData.R
\name{impute_sem}
\alias{impute_sem}
\alias{impute_sem,mids-method}
\title{Create SemImputedData Object}
\usage{
impute_sem(data, method, conf.int, conf.level)

\S4method{impute_sem}{mids}(data, method = "lavaan", conf.int = FALSE, conf.level = 0.95)
}
\arguments{
\item{data}{A \code{mids} object from the \code{mice} package containing multiply imputed datasets.}

\item{method}{A character string specifying the SEM analysis method.
Valid options are "lavaan" or "OpenMx". Defaults to "lavaan".}

\item{conf.int}{A logical value indicating whether confidence intervals are
included in the SEM results. Defaults to \code{FALSE}.}

\item{conf.level}{A numeric value specifying the confidence level for
confidence intervals, which must be between 0 and 1. Defaults to 0.95.
If \code{conf.int} is \code{FALSE}, this argument is ignored.}
}
\value{
An object of class SemImputedData. See \link{SemImputedData} for the details of the slots.
}
\description{
Constructs a new \code{SemImputedData} object for structural equation modeling (SEM) analysis
on multiply imputed datasets. This function ensures that the provided data is a \link[mice:mids-class]{mice::mids}
object from the \code{mice} package and that the specified SEM analysis method is supported.
}
\details{
All the arguments \code{data}, \code{method}, \code{conf.int}, and \code{conf.level} are used to specify the SEM analysis. \code{impute_sem} is a constructor function for \code{SemImputedData} class. These methods are used as constructors for the \code{SemImputedData} class.
}
\examples{
\dontrun{
data("HolzingerSwineford1939", package = "lavaan")
df_complete <- na.omit(HolzingerSwineford1939)
amp <- mice::ampute(df_complete, prop = 0.2, mech = "MAR")
imputed_data <- mice::mice(amp$amp, m = 3, maxit = 3, seed = 12345, printFlag = FALSE)
sem_data <- impute_sem(data = imputed_data, method = "lavaan")
str(sem_data)
}
}
\seealso{
\link{SemImputedData} \link[mice:mids-class]{mice::mids} \link{lavaan} \link{OpenMx}
}
