% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fit.R
\name{is_fit}
\alias{is_fit}
\alias{is_fit,MxModel-method}
\title{Determine If a SEM Model Has Been Fitted}
\usage{
is_fit(model,...)

\S4method{is_fit}{MxModel}(model, ...)
}
\arguments{
\item{model}{A model object, either of class \code{lavaan} or \code{MxModel}, representing
the SEM model to be checked.}

\item{...}{Additional arguments affecting the method dispatched
(currently not used but available for future extensions).}
}
\value{
A logical value: \code{TRUE} if the model has been fitted, and \code{FALSE} otherwise.
}
\description{
This function checks whether a structural equation modeling (SEM) model,
represented by either a \code{lavaan} object or an \code{MxModel} object from the \code{OpenMx}
package, has been fitted. The function offers a convenient way to programmatically
verify if the model fitting step has been executed for a given model object.
}
\examples{
\dontrun{
# Assuming 'lav_model' is a lavaan model object
lav_model_fit <- is_fit(lav_model)

# Assuming 'mx_model' is an OpenMx model object
mx_model_fit <- is_fit(mx_model)

# Checking the output
print(lav_model_fit)
print(mx_model_fit)
}
}
\seealso{
\link{lavaan}, \link{MxModel}
}
