% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lav_syntax.R
\name{is_lav_syntax}
\alias{is_lav_syntax}
\alias{is_lavaan}
\title{Function to check if lavaan model syntax is valid}
\usage{
is_lav_syntax(model)
}
\arguments{
\item{model}{A character string representing the lavaan model to be fitted.}
}
\value{
A logical value indicating whether the model syntax is valid.
}
\description{
This function checks if the lavaan model syntax is valid. It does so by
attempting to parse and fit the model in a safe environment. If the model
syntax is invalid, the function will return an error message.
}
\examples{
bad_model <- "y ~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8 + x9"
data(HolzingerSwineford1939, package = "lavaan")
is_lav_syntax(bad_model, HolzingerSwineford1939)
good_model <- "visual =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed =~ x7 + x8 + x9
visual ~ speed
textual ~ speed"
is_lav_syntax(good_model, HolzingerSwineford1939)
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
