% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pd.R
\name{is_pd}
\alias{is_pd}
\title{Checks if a matrix object is positive definite}
\usage{
is_pd(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{A numeric matrix.}
}
\value{
Returns \code{TRUE} if the matrix is positive definite, \code{FALSE} otherwise.
}
\description{
Determines if a matrix is positive definite (all eigenvalues
are strictly positive) by attempting Cholesky decomposition.
}
\examples{

# Example of a positive definite matrix
A <- matrix(c(1, 2, 2, 4), nrow = 2)
is_pd(A) # Should return TRUE
}
