% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{is_valid_for_computation}
\alias{is_valid_for_computation}
\title{Method to check if ProductNormal object is properly specified for computation}
\usage{
is_valid_for_computation(object)
}
\arguments{
\item{object}{A ProductNormal object}
}
\value{
TRUE if object is valid for computation methods
}
\description{
Method to check if ProductNormal object is properly specified for computation
}
