% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{MBCOResult}
\alias{MBCOResult}
\title{MBCO Result Class}
\usage{
MBCOResult(
  statistic = integer(0),
  df = integer(0),
  p_value = integer(0),
  type = character(0)
)
}
\arguments{
\item{statistic}{Numeric test statistic value.}

\item{df}{Numeric degrees of freedom.}

\item{p_value}{Numeric p-value.}

\item{type}{Character string indicating the type of test.}
}
\description{
A class representing the results of a Model-Based Constrained Optimization (MBCO) test.
}
