% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medci_prototype.R
\name{medci_prototype}
\alias{medci_prototype}
\title{Confidence Interval for the Mediated Effect (PROTOTYPE S7 Wrapper)}
\usage{
medci_prototype(
  mu.x,
  mu.y,
  se.x,
  se.y,
  rho = 0,
  alpha = 0.05,
  type = "dop",
  plot = FALSE,
  plotCI = FALSE,
  n.mc = 1e+05,
  ...
)
}
\arguments{
\item{mu.x}{mean of \eqn{x}}

\item{mu.y}{mean of \eqn{y}}

\item{se.x}{standard error (deviation) of \eqn{x}}

\item{se.y}{standard error (deviation) of \eqn{y}}

\item{rho}{correlation between \eqn{x} and \eqn{y}, where -1 <\code{rho} < 1.
The default value is 0.}

\item{alpha}{significance level for the confidence interval. The default
value is .05.}

\item{type}{method used to compute confidence interval. It takes on the
values \code{"dop"} (default), \code{"MC"}, \code{"asymp"} or \code{"all"}}

\item{plot}{when \code{TRUE}, plots the distribution of \code{n.mc} data
points from the distribution of product of two normal random variables
using the density estimates provided by the function \code{\link{density}}.
The default value is \code{FALSE}.}

\item{plotCI}{when \code{TRUE}, overlays a confidence interval with error
bars on the plot for the mediated effect. Note that to obtain the CI plot,
one must also specify \code{plot="TRUE"}. The default value is
\code{FALSE}.}

\item{n.mc}{when \code{type="MC"}, \code{n.mc} determines the sample size for
the Monte Carlo method. The default sample size is 1E5.}

\item{...}{additional arguments to be passed on to the function.}
}
\description{
Confidence Interval for the Mediated Effect (PROTOTYPE S7 Wrapper)
}
\keyword{internal}
