% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2m_par.R
\name{plot_2m_par}
\alias{plot_2m_par}
\title{Sensitivity plots for a two-mediator parallel model.}
\usage{
plot_2m_par(x, rho1_facet = c(-0.5, -0.3, -0.1, 0.1, 0.3, 0.5),
  rho2_facet = c(-0.5, -0.3, -0.1, 0.1, 0.3, 0.5), xlim1 = c(0, 0.5),
  xlim2 = c(0, 0.5), ylim1 = NULL, ylim2 = NULL, nrow = NULL,
  ncol = 2, device = NULL)
}
\arguments{
\item{x}{output from from \code{\link{sensitivity_2m_par}} function}

\item{rho1_facet}{facet (panel) values for the confounder correlation between Mediator 1 and the outcome variable}

\item{rho2_facet}{facet (panel) confounder correlation between Mediator 2 and the outcome variable}

\item{xlim1}{x-axis limits for the plot of indirect effect through  Mediator 1}

\item{xlim2}{x-axis limits for the plot of indirect effect through  Mediator 2}

\item{ylim1}{y-axis limits for the plot of indirect effect through  Mediator 1}

\item{ylim2}{y-axis limits for the plot of indirect effect through  Mediator 2}

\item{nrow}{the number of rows for matrix of plots}

\item{ncol}{the number of columns for the matrix of plots}

\item{device}{type of file to save the plot: 'pdf', 'png', etc}
}
\description{
Sensitivity plots for a two-mediator parallel model.
}
\seealso{
\code{\link{sensitivity_2m_par}}
}
