% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{pool}
\alias{pool}
\title{Pool Method}
\usage{
pool(x, ...)
}
\arguments{
\item{x}{Object to be pooled.}

\item{...}{Additional arguments to methods.}
}
\description{
Description of what the pool function is supposed to do.
}
