% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_sem.R
\name{pool_sem}
\alias{pool_sem}
\alias{pool_sem,SemResults-method}
\title{Pool SEM Analysis Results}
\usage{
pool_sem(object, conf.int, conf.level, ...)

\S4method{pool_sem}{SemResults}(object, conf.int, conf.level, ...)
}
\arguments{
\item{object}{A \code{SemResults} object containing the SEM analysis results from
multiple imputations, along with the SEM method used (\code{lavaan} or \code{OpenMx}),
and confidence interval settings.}

\item{conf.int}{Logical; indicates whether confidence intervals should be
calculated for the pooled estimates. Defaults to the value specified in the
\code{SemResults} object.}

\item{conf.level}{Numeric; specifies the confidence level to use for the
confidence interval calculations, if \code{conf.int} is \code{TRUE}. Defaults to the
value specified in the \code{SemResults} object.}

\item{...}{Additional arguments for future use or extensions.}
}
\value{
A summary object with pooled SEM analysis results, which might include
estimates, standard errors, confidence intervals, etc., depending on the method.

A \code{data.frame} containing the pooled results of the SEM analyses. The
column names adhere to tidy conventions and include the following columns:
\itemize{
\item \code{term}: The name of the parameter being estimated.
\item \code{estimate}: The pooled estimate of the parameter.
\item \code{std.error}: The pooled standard error of the estimate.
\item \code{statistic}: The pooled test statistic (e.g., z-value, t-value).
\item \code{p.value}: The pooled p-value for the test statistic.
\item \code{conf.low}: The lower bound of the confidence interval for the estimate.
\item \code{conf.high}: The upper bound of the confidence interval for the estimate.
}
}
\description{
A generic function to pool SEM analysis results from multiple datasets or imputations,
offering a unified interface for summarizing findings across different SEM models
or imputations. Supports custom implementations for various result types, including
those from \code{lavaan} and \code{OpenMx}.

This function pools the results of structural equation modeling (SEM) analyses
performed on multiple imputed datasets. It supports pooling for models analyzed
with either the \code{lavaan} or \code{OpenMx} package. The function extracts and pools
relevant statistics (e.g., estimates, standard errors) across all imputations,
considering the specified confidence interval settings.
}
\details{
Refer to method-specific documentation for details on pooling process and assumptions.
}
\examples{
\dontrun{
pooled_lavaan <- pool_sem(lavaan_results) # For `lavaan` results
pooled_openmx <- pool_sem(mx_results) # For `OpenMx` results
}
\dontrun{
# Assuming `sem_results` is a SemResults object with lavaan model fits:
pooled_results <- pool_sem(sem_results)

# If you want to calculate and include confidence intervals at a 95\% level:
pooled_results_ci <- pool_sem(sem_results, conf.int = TRUE, conf.level = 0.95)
}

}
\seealso{
\link{lavaan}, \link{OpenMx}

\code{\link[lavaan]{lavaan}}, \code{\link[OpenMx]{mxModel}}
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
