% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_semMice.R
\name{pool.semMice}
\alias{pool.semMice}
\alias{pool_semMice}
\title{Pooling function for 'mira' objects of class 'semMice'}
\usage{
pool.semMice(object, ...)
}
\arguments{
\item{object}{An object of class 'mira' representing the results of fitting a SEM model to multiply imputed datasets}

\item{...}{Additional arguments to be passed to the pooling function}
}
\value{
A pooled 'mira' object representing the pooled results of fitting a SEM model to multiply imputed datasets
}
\description{
Pooling function for 'mira' objects of class 'semMice'
}
\examples{
\dontrun{
# library(mice)
# library(lavaan)
# Load Holzinger and Swineford (1939) dataset
data("HolzingerSwineford1939", package = "lavaan")
# Introduce missing data
df_complete <- na.omit(HolzingerSwineford1939)
amp <- mice::ampute(df_complete, prop = 0.2, mech = "MAR")
data_with_missing <- amp$amp

# Perform multiple imputation
imputed_data <- mice::mice(data_with_missing, m = 3, maxit = 5, seed = 12345, printFlag = FALSE)

# fit the Holzinger and Swineford (1939) example model
HS_model <- "visual  =~ x1 + x2 + x3
            textual =~ x4 + x5 + x6
            speed   =~ x7 + x8 + x9 "

# Fit the SEM model without running the model
fit_HS <- lavaan::sem(HS_model, data = data_with_missing, do.fit = FALSE)

# Fit the SEM model without pooling to each imputed dataset
fit_list <- lav_mice(HS_model, imputed_data)
pooled_fit <- pool(fit_list)

  }

}
\seealso{
\code{\link{pool}}
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
