% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledSEMResults.R
\docType{class}
\name{PooledSEMResults}
\alias{PooledSEMResults}
\alias{PooledSEMResults-class}
\title{Pooled SEM Analysis Results Class}
\description{
An S4 class to represent pooled results from SEM analysis across multiple imputations or datasets.
It contains pooled estimates, standard errors, test statistics, p-values, and confidence intervals
for each parameter estimated across multiple imputations.
}
\section{Slots}{

\describe{
\item{\code{tidy_table}}{data.frame A data frame containing the pooled results of the SEM analyses. The column names adhere to tidy conventions and include the following columns:
\itemize{
\item \code{term}: The name of the parameter being estimated.
\item \code{estimate}: The pooled estimate of the parameter.
\item \code{std_error}: The pooled standard error of the estimate.
\item \code{p_value}: The pooled p-value for the test statistic.
\item \code{conf_low}: The lower bound of the confidence interval for the estimate.
\item \code{conf_high}: The upper bound of the confidence interval for the estimate.
}}

\item{\code{cov_total}}{matrix The pooled total covariance matrix of the parameter estimates.}

\item{\code{cov_between}}{matrix The pooled between-imputation covariance matrix of the parameter estimates.}

\item{\code{cov_within}}{matrix The pooled within-imputation covariance matrix of the parameter estimates.}

\item{\code{method}}{character The method used for SEM analysis ('lavaan' or 'OpenMx').}

\item{\code{conf_int}}{logical Whether to calculate confidence intervals for the pooled estimates. default is \code{FALSE}.}

\item{\code{conf_level}}{numeric The confidence level used in the interval calculation. default is \code{0.95}.}
}}

\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
