% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{print}
\alias{print}
\title{Print method}
\usage{
print(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for printing objects.
}
