% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{ProductNormal_from_lavaan}
\alias{ProductNormal_from_lavaan}
\title{Utility function to create ProductNormal from lavaan parameter estimates}
\usage{
ProductNormal_from_lavaan(lavaan_model, param_names)
}
\arguments{
\item{lavaan_model}{A fitted lavaan model object}

\item{param_names}{Names of parameters to include in the product (e.g., c("a", "b"))}
}
\value{
A ProductNormal object
}
\description{
Utility function to create ProductNormal from lavaan parameter estimates
}
