% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qprodnormal.R
\name{qprodnormal}
\alias{qprodnormal}
\title{Quantile for the Distribution of Product of Two Normal Variables}
\usage{
qprodnormal(
  p,
  mu.x,
  mu.y,
  se.x,
  se.y,
  rho = 0,
  lower.tail = TRUE,
  type = "dop",
  n.mc = 1e+05
)
}
\arguments{
\item{p}{probability}

\item{mu.x}{mean of \eqn{x}}

\item{mu.y}{mean of \eqn{y}}

\item{se.x}{standard error (deviation) of \eqn{x}}

\item{se.y}{standard error (deviation) of \eqn{y}}

\item{rho}{correlation between \eqn{x} and \eqn{y}, where -1 <\code{rho} < 1.
The default value is 0.}

\item{lower.tail}{logical; if \code{TRUE} (default), the probability is
\eqn{P[X*Y < q]}; otherwise, \eqn{P[X*Y > q]}}

\item{type}{method used to compute confidence interval. It takes on the
values \code{"dop"} (default), \code{"MC"}, \code{"asymp"} or \code{"all"}}

\item{n.mc}{when \code{type="MC"}, \code{n.mc} determines the sample size for
the Monte Carlo method. The default sample size is 1E5.}
}
\value{
An object of the type \code{\link{list}} that contains the
following values: \item{q}{quantile corresponding to probability \code{p}}
\item{error}{estimate of the absolute error}
}
\description{
Generates quantiles for the distribution of product of two normal random
variables
}
\details{
This function returns a quantile and the associated error
(accuracy) corresponding the requested percentile (probability) \code{p} of
the distribution of product of mediated effect (product of two normal
random variables). To obtain a quantile using a specific method, the
argument \code{type} should be specified. The default method is
\code{type="dop"}, which uses the method described by Meeker and Escobar
(1994) to evaluate the CDF of the distributionof product of two normal
variables. \code{type="MC"} uses the Monte Carlo approach (Tofighi &
MacKinnon, 2011). \code{type="all"} prints quantiles using all three
options. For the method \code{type="dop"}, the error is the modulus of
absolute error for the numerical integration (for more information see
Meeker and Escobar, 1994). For \code{type="MC"}, the error refers to the
Monte Carlo error.
}
\examples{
##lower tail
qprodnormal(p=.1, mu.x=.5, mu.y=.3, se.x=1, se.y=1, rho=0,
lower.tail = TRUE, type="all")
##upper tail
qprodnormal(p=.1, mu.x=.5, mu.y=.3, se.x=1, se.y=1, rho=0,
lower.tail = FALSE, type="all")
}
\references{
Tofighi, D. and MacKinnon, D. P. (2011). RMediation: An R
package for mediation analysis confidence intervals. \emph{Behavior
Research Methods}, \bold{43}, 692--700. doi:10.3758/s13428-011-0076-x
}
\seealso{
\code{\link{medci}} \code{\link{RMediation-package}}
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
