% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{quantile}
\alias{quantile}
\title{Quantile Function}
\usage{
quantile(object, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for computing quantiles.
}
