% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mplus.R
\name{run_mplus}
\alias{run_mplus}
\title{Run Mplus input files within a directory.}
\usage{
run_mplus(path = tempdir(), cores = parallel::detectCores() - 1,
  outPath = tempdir(), logPath = tempdir(), pattern = NULL)
}
\arguments{
\item{path}{A path to the mplus files. Default is \code{\link{tempdir}}.}

\item{cores}{The number of cores. Default is \code{\link{detectCores}-1}}

\item{outPath}{output path}

\item{logPath}{log files path}

\item{pattern}{pattern}
}
\value{
Does not return a value.
}
\description{
Run Mplus input files within a directory.
}
