% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_mice.R
\name{sem_mice}
\alias{sem_mice}
\title{Fit a structural equation model to multiply imputed data}
\usage{
sem_mice(model, mids, ...)
}
\arguments{
\item{model}{Either a character string representing the structural equation
model to be fitted or a pre-fitted \code{lavaan} or \code{OpenMx} model object.}

\item{mids}{A \code{mice::mids} object from the \code{mice} package.}

\item{...}{Additional arguments to be passed to \verb{[lavaan::sem]} or
\verb{[OpenMx::mxRun]}.}
}
\value{
A list of \code{lavaan} or \code{OpenMx} model fits, one for each imputed
dataset. The class of the returned object is \code{lav} for lavaan /\code{mx} for \code{MxModel}, \code{semlist}, and \code{list}.
}
\description{
This function fits a structural equation model to multiply imputed data. It
is an extension for the \code{\link[lavaan:sem]{lavaan::sem()}} and \link[OpenMx:MxModel-class]{OpenMx::MxModel} functions to
handle \link[mice:mids-class]{mice::mids} objects from the \link{mice} package. It allows for both a
structural equation model syntax as a character string or a pre-fitted
\code{lavaan} or \code{OpenMx} model object. This function is a wrapper for the
\link{lav_mice} and \link{mx_mice} functions.
}
\examples{
\dontrun{
# library(mice)
# library(lavaan)
# Load Holzinger and Swineford (1939) dataset
data("HolzingerSwineford1939", package = "lavaan")
# Introduce missing data
df_complete <- na.omit(HolzingerSwineford1939)
amp <- mice::ampute(df_complete, prop = 0.2, mech = "MAR")
data_with_missing <- amp$amp

# Perform multiple imputation
imputed_data <- mice::mice(data_with_missing, m = 3, maxit = 5, seed = 12345, printFlag = FALSE)

# fit the Holzinger and Swineford (1939) example model
HS_model <- "visual  =~ x1 + x2 + x3
            textual =~ x4 + x5 + x6
           speed   =~ x7 + x8 + x9 "
# Fit the SEM model without running the model
fit_HS <- lavaan::sem(HS_model, data = data_with_missing, do.fit = FALSE)
# Fit the SEM model without pooling to each imputed dataset
fit_list1 <- sem_mice(HS_model, imputed_data)
# 'fit_list1' now contains a list of lavaan objects, one for each imputed dataset
# Fit the SEM model without pooling to each imputed dataset using a pre-fitted model object
fit_list2 <- sem_mice(fit_HS, imputed_data)

# Assuming mx_mice is correctly defined in your environment
fits <- mx_mice(model, imputed_data)
summary(fits[[1]])
}
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
