% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemImputedData.R
\docType{class}
\name{SemImputedData}
\alias{SemImputedData}
\alias{SemImputedData-class}
\title{SemImputedData Class}
\description{
An S4 class to hold multiply imputed datasets for structural equation modeling (SEM) analysis.
It facilitates working with imputed data from the \code{mice} package and supports SEM analysis
using either the \code{lavaan} or \code{OpenMx} packages.
}
\section{Slots}{

\describe{
\item{\code{data}}{An object of class \code{mids} from the \code{mice} package, representing multiply imputed datasets.}

\item{\code{method}}{A character string indicating the SEM package to be used for analysis.
Valid options are "lavaan" or "OpenMx". Defaults to "lavaan".}

\item{\code{conf.int}}{A logical value indicating whether confidence intervals are
included in the SEM results. Defaults to \code{FALSE}.}

\item{\code{conf.level}}{A numeric value specifying the confidence level for
confidence intervals, which must be between 0 and 1. Defaults to 0.95.}
}}

\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
