% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemResults.R
\docType{class}
\name{SemResults}
\alias{SemResults}
\alias{SemResults-class}
\title{SemResults Class}
\description{
An S4 class for storing the results of structural equation modeling analysis
performed on multiply imputed datasets. This class can accommodate results
from \link{lavaan}, \link{OpenMx}, or potentially other SEM software, providing
a unified interface to access and manipulate SEM results.
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of SEM model fits, where each element represents the
results from one imputed dataset. The specific content and structure of
each element can vary based on the SEM method used (e.g., objects of
class \code{lavaan} or \code{MxModel}). The \code{results} slot is the primary storage
location for SEM results, and it is expected to be a list of length equal
to the number of imputed datasets.}

\item{\code{method}}{A character string specifying the SEM package used for the
analysis, with valid options being "lavaan" for \code{lavaan} model fits
or "OpenMx" for \code{OpenMx} model fits. Defaults to "lavaan".}

\item{\code{conf.int}}{A logical value indicating whether confidence intervals are
included in the SEM results. Defaults to \code{FALSE}.}

\item{\code{conf.level}}{A numeric value specifying the confidence level for
confidence intervals, which must be between 0 and 1. Defaults to 0.95.}
}}

\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
