% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_2m_par.R
\name{sensitivity_2m_par}
\alias{sensitivity_2m_par}
\title{Sensitivity analysis for a two-mediator parallel model.}
\usage{
sensitivity_2m_par(input_file = "", rho1 = c(-0.5, 0.5),
  rho2 = c(-0.5, 0.5), path = tempdir(), alpha = 0.05, lab = list(x
  = "X", m1 = "I", m2 = "S", y = "Y", ind1 = "ind1", ind2 = "ind2"))
}
\arguments{
\item{input_file}{name and path to Mplus input file. This file needs to follow specific instruction. See below for more detail.}

\item{rho1}{confounder correlation between Mediator 1 and the outcome variable}

\item{rho2}{confounder correlation between Mediator 2 and the outcome variable}

\item{path}{directory and its path to save all the files generated when condcuting sensitivity analysis. Default is \code{\link{tempdir}}, in which the output files will be saved and deleted the R session ends.}

\item{alpha}{signficance level}

\item{lab}{lables used in the imput files to describe variables and indirect effect in the mplus file specfied in \code{input_file}.}
}
\description{
Sensitivity analysis for a two-mediator parallel model.
}
