% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemImputedData.R
\name{set_sem}
\alias{set_sem}
\alias{set_sem-methods}
\alias{set_sem,mids-method}
\title{Create SemImputedData Object}
\usage{
set_sem(data, method = "lavaan", conf_int = FALSE, conf_level 0.95)

\S4method{set_sem}{mids}(data, model, conf_int = FALSE, conf_level = 0.95)
}
\arguments{
\item{data}{A \code{mids} object from the \code{mice} package containing multiply imputed datasets.}

\item{model}{A \code{lavaan} or \code{OpenMx} model syntax to be used for SEM analysis. For \code{lavaan} models, the syntax should be a character string as described in \link[lavaan:model.syntax]{lavaan::model.syntax}. For \code{OpenMx} models, the syntax should be an \link{mxModel} object with or without \code{\link[=mxData]{mxData()}} specified; that is, \code{mxModel} syntax can be without data specified. In addition, both \code{lavaan} and \code{OpenMx} models can be a fitted model object in the respective package.}

\item{conf_int}{A logical value indicating whether confidence intervals are
included in the SEM results. Defaults to \code{FALSE}.}

\item{conf_level}{A numeric value specifying the confidence level for
confidence intervals, which must be between 0 and 1. Defaults to 0.95.
If \code{conf_int} is \code{FALSE}, this argument is ignored.}
}
\value{
An object of class SemImputedData. See \link{SemImputedData} for the details of the slots.
}
\description{
Constructs a new \code{SemImputedData} object for structural equation modeling (SEM) analysis
on multiply imputed datasets. This function ensures that the provided data is a \link[mice:mids-class]{mice::mids}
object from the \code{mice} package and that the specified SEM analysis method is supported.
}
\details{
All the arguments \code{data}, \code{method}, \code{conf_int}, and \code{conf_level} are used to specify the SEM analysis. \code{set_sem} is a constructor function for \code{SemImputedData} class. These methods are used as constructors for the \code{SemImputedData} class.
}
\examples{
\dontrun{
data("HolzingerSwineford1939", package = "lavaan")
df_complete <- na.omit(HolzingerSwineford1939)
amp <- mice::ampute(df_complete, prop = 0.2, mech = "MAR")
imputed_data <- mice::mice(amp$amp, m = 3, maxit = 3, seed = 12345, printFlag = FALSE)
sem_data <- set_sem(data = imputed_data, method = "lavaan")
str(sem_data)
}
}
\seealso{
\link{SemImputedData} \link[mice:mids-class]{mice::mids} \link{lavaan} \link{OpenMx}
}
