% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,MediationCI-method}
\alias{show,MediationCI-method}
\title{Show Summary of a MediationCI Object}
\usage{
\S4method{show}{MediationCI}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{MediationCI}.}
}
\value{
Invisible. The function is called for its side effect of printing
the summary.
}
\description{
Displays a concise summary of a \code{MediationCI} object, including the method
used for confidence interval computation, confidence level, point estimate of the
mediated effect, confidence interval bounds, and the type of SEM model attached.
}
\examples{
\dontrun{
# Assuming `mediationResult` is an existing MediationCI object
show(mediationResult)
}

}
\seealso{
\link{show}, \linkS4class{MediationCI}
}
\keyword{internal}
