% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{show}
\alias{show}
\title{Show method}
\usage{
show(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for showing objects.
}
