% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,SemImputedData-method}
\alias{summary,SemImputedData-method}
\title{Summary Method for SemImputedData Objects}
\usage{
\S4method{summary}{SemImputedData}(object, ...)
}
\arguments{
\item{object}{An object of class SemImputedData.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A textual summary of the SemImputedData object.
}
\description{
Provides a comprehensive summary of an SemImputedData object, including the SEM method used,
the number of imputations, basic information about the imputed data, and summaries of the
original data and fitted model.
}
\examples{
\dontrun{
# Assuming `sem_imputed_data` is an SemImputedData object
summary(sem_imputed_data)
}
}
