% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{summary}
\alias{summary}
\title{Summary method}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for summarizing objects.
}
