% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_indirect.R
\name{test_indirect}
\alias{test_indirect}
\alias{test_indirect,lavaan-method}
\alias{test_indirect,MxModel-method}
\title{Performing Tests of Indirect Effects}
\usage{
test_indirect(model, method, conf.level, estimate, conf.low, conf.high, se, ...)

\S4method{test_indirect}{lavaan}(
  model,
  method = "dop",
  conf.level = 0.95,
  estimate,
  conf.low,
  conf.high,
  se,
  ...
)

\S4method{test_indirect}{MxModel}(
  model,
  method = "dop",
  conf.level = 0.95,
  estimate,
  conf.low,
  conf.high,
  se,
  ...
)
}
\arguments{
\item{model}{An \code{OpenMx} model object. This can be a fitted model object or a model specification.}

\item{method}{Character, the method used to compute the confidence interval. The default is "dop" for the distribution of the product method. Other options include "MC" for the Monte Carlo method.}

\item{conf.level}{Numeric, the confidence level used in the interval calculation. The default is 0.95.}

\item{estimate}{Numeric, the point estimate of the mediated effect.}

\item{conf.low}{Numeric, the lower bound of the confidence interval.}

\item{conf.high}{Numeric, the upper bound of the confidence interval.}

\item{se}{Numeric, the standard error of the mediated effect estimate.}

\item{...}{Additional arguments to customize the object creation process.}
}
\value{
An object of class \code{MediationCI}.
}
\description{
Perform tests and compute confidence intervals for indirect effects.

Method for \code{lavaan} model objects. This method specifically handles models fitted with the \code{lavaan} package.

Method for \code{OpenMx} model objects. This method specifically handles models fitted with the \code{OpenMx} package.
}
\details{
The \code{test_indirect} function creates instances of the \code{MediationCI} class, encapsulating confidence intervals for mediation analysis results. The \code{MediationCI} class is an S4 class that stores the results of mediation analysis, including the mediated effect estimate, confidence interval, and additional statistics.
}
\note{
This method treats the \code{model} argument as a \code{lavaan} model object and may handle additional arguments unique to \code{lavaan} analysis.

This method treats the \code{model} argument as an \code{OpenMx} model object and may handle additional arguments unique to \code{OpenMx} analysis.
}
\examples{
\dontrun{
my_fitted_model <- ... # Assume fitting a model with OpenMx or lavaan
my_mediation_ci <- test_indirect(model = my_fitted_model, conf.level = 0.95)
}
}
\seealso{
\linkS4class{MediationCI}
}
\keyword{internal}
