% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMediation.R
\name{tidy}
\alias{tidy}
\title{Tidy generic function}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{An object to tidy}

\item{...}{Additional arguments passed to methods}
}
\value{
A tibble or data.frame with tidy output
}
\description{
Tidy generic function
}
