% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_ProductNormal}
\alias{validate_ProductNormal}
\title{Additional validation for ProductNormal objects}
\usage{
validate_ProductNormal(object, verbose = FALSE)
}
\arguments{
\item{object}{A ProductNormal object}

\item{verbose}{Whether to show detailed validation messages}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Additional validation for ProductNormal objects
}
