% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{cdf}
\alias{cdf}
\title{Cumulative Distribution Function}
\usage{
cdf(object, ...)
}
\arguments{
\item{object}{A distribution object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for computing cumulative distribution function.
}
