% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{ci}
\alias{ci}
\title{Confidence Interval}
\usage{
ci(mu, ...)
}
\arguments{
\item{mu}{A distribution object or numeric vector of means.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for computing confidence intervals.
}
