% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sem.R
\name{impute_sem}
\alias{impute_sem}
\alias{impute_sem,mids-method}
\title{Create SemImputedData Object}
\usage{
impute_sem(data, method)

\S4method{impute_sem}{mids}(data, method = "lavaan")
}
\arguments{
\item{data}{A \code{mids} object from the \code{mice} package containing multiply imputed datasets.}

\item{method}{A character string specifying the SEM analysis method.
Valid options are "lavaan" or "OpenMx". Defaults to "lavaan".}
}
\value{
An object of class \code{SemImputedData}.
}
\description{
Constructs a new \code{SemImputedData} object for structural equation modeling (SEM) analysis
on multiply imputed datasets. This function ensures that the provided data is a \link[mice:mids-class]{mice::mids}
object from the \code{mice} package and that the specified SEM analysis method is supported.
}
\examples{
\dontrun{
data("HolzingerSwineford1939", package = "lavaan")
df_complete <- na.omit(HolzingerSwineford1939)
amp <- mice::ampute(df_complete, prop = 0.2, mech = "MAR")
imputed_data <- mice::mice(amp$amp, m = 3, maxit = 3, seed = 12345, printFlag = FALSE)
sem_data <- impute_sem(data = imputed_data, method = "lavaan")
str(sem_data)
}
}
\seealso{
\link[mice:mids-class]{mice::mids} \link{lavaan} \link{OpenMx}
}
