% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medci.R
\name{medci}
\alias{medci}
\title{Confidence Interval for the Mediated Effect}
\usage{
medci(
  mu.x,
  mu.y,
  se.x,
  se.y,
  rho = 0,
  alpha = 0.05,
  type = "dop",
  n.mc = 1e+05,
  ...
)
}
\arguments{
\item{mu.x}{Mean of \eqn{x}.}

\item{mu.y}{Mean of \eqn{y}.}

\item{se.x}{Standard error of \eqn{x}.}

\item{se.y}{Standard error of \eqn{y}.}

\item{rho}{Correlation between \eqn{x} and \eqn{y} (-1 < \code{rho} < 1).}

\item{alpha}{Significance level for the CI (default: 0.05).}

\item{type}{Method for CI: "dop" (default), "MC", "asymp", or "all".}

\item{n.mc}{Sample size for Monte Carlo method when \code{type="MC"} (default: 1e5).}

\item{...}{Additional arguments for future development.}
}
\value{
A list containing CI, estimates, and additional information based on the \code{type}.
}
\description{
Computes confidence intervals for the mediated effect, supporting various methods.
}
\examples{
res <- medci(mu.x = .2, mu.y = .4, se.x = 1, se.y = 1, type = "dop")
}
\references{
Craig, C. C. (1936). On the frequency function of \eqn{xy}.
\emph{The Annals of Mathematical Statistics}, \bold{7}, 1--15.

MacKinnon, D. P., Fritz, M. S., Williams, J., and Lockwood, C. M. (2007).
Distribution of the product confidence limits for the indirect effect:
Program PRODCLIN. \emph{Behavior Research Methods}, \bold{39}, 384--389.

Meeker, W. and Escobar, L. (1994). An algorithm to compute the CDF of the
product of two normal random variables. \emph{Communications in Statistics:
Simulation and Computation}, \bold{23}, 271--280.

Tofighi, D. and MacKinnon, D. P. (2011). RMediation: An R package for mediation analysis
confidence intervals. \emph{Behavior Research Methods}, \bold{43},
692--700. \doi{doi:10.3758/s13428-011-0076-x}
}
\seealso{
\code{\link{lavaan}}, \code{\link[OpenMx]{mxModel}}
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
\keyword{confidence-interval}
\keyword{mediation}
