% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemResults.R
\name{pool_sem}
\alias{pool_sem}
\alias{pool_sem,SemResults-method}
\title{Pool SEM Analysis Results}
\usage{
pool_sem(object)

\S4method{pool_sem}{SemResults}(object)
}
\arguments{
\item{object}{\code{SemResults} object with SEM analysis results.}
}
\value{
\code{PooledSEMResults} object containing pooled SEM analysis results.

A \code{data.frame} containing the pooled results of the SEM analyses. The column names adhere to tidy conventions and include the following columns:
\itemize{
\item \code{term}: The name of the parameter being estimated.
\item \code{estimate}: The pooled estimate of the parameter.
\item \code{std_error}: The pooled standard error of the estimate.
\item \code{statistic}: The pooled test statistic (e.g., z-value, t-value).
\item \code{p_value}: The pooled p-value for the test statistic.
\item \code{conf_low}: The lower bound of the confidence interval for the estimate.
\item \code{conf_high}: The upper bound of the confidence interval for the estimate.
}
}
\description{
\code{pool_sem} pools SEM analysis results, supporting \code{lavaan} and \code{OpenMx} models.
It calculates pooled estimates, standard errors, confidence intervals, and more.

This function pools the results of structural equation modeling (SEM)
analyses performed on multiple imputed datasets.
It supports pooling for models analyzed with either the \code{lavaan}
or \code{OpenMx} package.  The function extracts and pools
relevant statistics (e.g., estimates, standard errors) across
all imputations, considering the specified confidence interval settings.
}
\details{
A generic function to pool SEM analysis results from multiple datasets or imputations.

Refer to method-specific documentation for details on pooling process and assumptions.
}
\examples{
\dontrun{
# Assuming `sem_results` is a SemResults object with lavaan model fits:
pooled_results <- pool_sem(sem_results)

# If you want to calculate and include confidence intervals at a 95\% level:
pooled_results_ci <- pool_sem(sem_results, conf_int = TRUE, conf_level = 0.95)
}

}
\seealso{
\link{lavaan}, \link{OpenMx}
}
\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
