% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PooledSEMResults.R
\docType{class}
\name{PooledSEMResults}
\alias{PooledSEMResults}
\alias{PooledSEMResults-class}
\title{Pooled SEM Analysis Results Class}
\description{
An S4 class to represent pooled results from SEM analysis across multiple imputations or datasets.
It contains pooled estimates, standard errors, test statistics, p-values, and confidence intervals
for each parameter estimated across multiple imputations.
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame A data frame containing the pooled results of the SEM analyses. The column names adhere to tidy conventions and include the following columns:
\itemize{
\item \code{term}: The name of the parameter being estimated.
\item \code{estimate}: The pooled estimate of the parameter.
\item \code{std.error}: The pooled standard error of the estimate.
\item \code{statistic}: The pooled test statistic (e.g., z-value, t-value).
\item \code{p.value}: The pooled p-value for the test statistic.
\item \code{conf.low}: The lower bound of the confidence interval for the estimate.
\item \code{conf.high}: The upper bound of the confidence interval for the estimate.
}}

\item{\code{method}}{character The method used for SEM analysis ('lavaan' or 'OpenMx').}
}}

\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
