% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{ProductNormal}
\alias{ProductNormal}
\title{ProductNormal Class}
\usage{
ProductNormal(mu = integer(0), Sigma = integer(0))
}
\arguments{
\item{mu}{Numeric vector of means.}

\item{Sigma}{Covariance matrix.}
}
\description{
Represents the distribution of the product of normal random variables.
}
