% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{ProductNormal3}
\alias{ProductNormal3}
\title{ProductNormal3 Class}
\usage{
ProductNormal3(mu = integer(0), Sigma = integer(0), method = character(0))
}
\arguments{
\item{mu}{Numeric vector of means of length 3.}

\item{Sigma}{3x3 covariance matrix.}

\item{method}{Integration method. Currently only \code{"hcubature"} is
supported.}
}
\description{
Represents the distribution of the product of three normal random variables.
Intended for sequential indirect effects of the form \code{a1 * a2 * b}.
}
