% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMediation-package.R
\docType{package}
\name{RMediation-package}
\alias{RMediation}
\alias{RMediation-package}
\title{RMediation: Mediation Analysis Confidence Intervals}
\description{
We provide functions to compute confidence intervals for a well-defined nonlinear function of the model parameters (e.g., product of k coefficients) in single--level and multilevel structural equation models. It also computes a chi-square test statistic for a function of indirect effects. 'Tofighi', D. and 'MacKinnon', D. P. (2011). 'RMediation' An R package for mediation analysis confidence intervals. Behavior Research Methods, 43, 692--700. \doi{10.3758/s13428-011-0076-x}. 'Tofighi', D. (2020). Bootstrap Model-Based Constrained Optimization Tests of Indirect Effects. Frontiers in Psychology, 10, 2989. \doi{10.3389/fpsyg.2019.02989}.
}
\author{
\strong{Maintainer}: Davood Tofighi \email{dtofighi@gmail.com} (\href{https://orcid.org/0000-0001-8523-7776}{ORCID})

}
\keyword{internal}
