% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMediation-package.R
\docType{package}
\name{RMediation-package}
\alias{RMediation}
\alias{RMediation-package}
\title{RMediation: Mediation Analysis Confidence Intervals}
\description{
We provide functions to compute confidence intervals for a well-defined nonlinear function of the model parameters (e.g., product of k coefficients) in single--level and multilevel structural equation models.
}
\keyword{internal}
