% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemResults.R
\docType{class}
\name{SemResults}
\alias{SemResults}
\title{SemResults Class}
\description{
An S4 class for storing the results of SEM analysis
performed on multiply imputed datasets. Supports \code{lavaan} and \code{OpenMx}.
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of SEM model fits for each imputed dataset.}

\item{\code{estimate_df}}{Data frame of parameter estimates and standard errors.}

\item{\code{coef_df}}{Data frame of coefficient estimates for each imputed dataset.}

\item{\code{cov_df}}{List of covariance matrices of coefficient estimates.}

\item{\code{method}}{SEM package used for analysis: 'lavaan' or 'OpenMx'.}

\item{\code{conf_int}}{Logical; if confidence intervals are included.}

\item{\code{conf_level}}{Confidence level for confidence intervals.}
}}

\author{
Davood Tofighi \email{dtofighi@gmail.com}
}
